Fixed warning with type conversion.

+++ Rogue-Robots/Runtime/src/Game/LuaInterfaces.cpp
@@ -123,7 +123,7 @@ void EntityInterface::GetPlayerStats(DOG::LuaContext* context)
 
 void EntityInterface::AddModel(LuaContext* context, entity e)
 {
-	EntityManager::Get().AddComponent<ModelComponent>(e, std::stoull(context->GetString()));
+	EntityManager::Get().AddComponent<ModelComponent>(e, static_cast<u32>(std::stoull(context->GetString())));
 }
 
 void EntityInterface::AddTransform(LuaContext* context, entity e)
